refactor: AcceptanceTest에 대한 중복된 메서드 제거
- 추상 클래스로 중복 제거
- createUser, login, getIdFromUrl, createPostWithoutImage, findPost
